requirements.txt
opencv-python~=4.9.0.80
flask~=2.2.5
roboflow~=1.1.27
requests~=2.31.0
pyserial~=3.5
ultralytics~=8.0.196